Se actualizo el modelo, form, view y templates para agregar y listar trabajadores
